requirements.txt
-e .
Django==4.1.5
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
pandas==1.5.2
python-dotenv==1.0.0
mysqlclient==2.1.1
python-dotenv==1.0.0